+++ Runtime/Input/IClickable.cs
@@ -4,8 +4,11 @@ namespace GameUtils
 {
     public interface IClickable
     {
-        void OnMouseEnter(Vector3 mousePosition);
-        void OnMouseExit();
-        void OnMouseClick(Vector3 hitPoint);
+        int Priority { get; }
+
+        //
+        void OnEnter(Vector3 mousePosition);
+        void OnExit();
+        void OnClick(Vector3 hitPoint);
     }
 }